--- a/index.js
+++ b/index.js
@@ -8,6 +8,7 @@ const defaultCacheKey = (...args) => {
 	if (args.length === 0) {
 		return '__defaultKey';
 	}
+
 	if (args.length === 1) {
 		const [firstArgument] = args;
 		if (