@sindresorhus 2018-08-27
Very nice! Thanks for doing this :)